pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mkdocs-dtinth"
version = "0.1.0"
description = "MkDocs with custom dtinth theme"
readme = "README.md"
requires-python = ">=3.8"
dependencies = ["mkdocs", "pyyaml", "mkdocs-material"]

[project.scripts]
mkdocs-dtinth = "mkdocs_dtinth.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["mkdocs_dtinth*"]

[tool.setuptools.package-data]
mkdocs_dtinth = ["theme/**/*"]

[project.entry-points.'mkdocs.themes']
dtinth = "mkdocs_dtinth.theme"
